working on add and update posts

+++ controllers/post.js
@@ -1,3 +1,4 @@
+import jwt from "jsonwebtoken"
 import {db} from "../db.js"
 export const getPosts=(req,res)=>{
     const query=req.query.cat ? "select * from posts where cat=?":"select * from posts";
@@ -9,7 +10,7 @@ export const getPosts=(req,res)=>{
 }
 export const getPost=(req,res)=>{
     const query="SELECT name, title, `desc`,p.img, p.cat, p.id AS postid, u.img AS userimg,date FROM blog.users u JOIN blog.posts p ON u.id=p.userid WHERE p.id=?";
-    console.log(req.params.id)
+    // console.log(req.params.id)
     db.query(query, [req.params.id],(err, data)=>{
         if(err)return res.status(500).json("Error to fetch post try to reload")
             console.log(data[0])
@@ -20,10 +21,30 @@ export const getPost=(req,res)=>{
 }
 export const addPost=(req,res)=>{
     const {title,desc,cat}=req.body
+    const token=req.cookie.auth__token;
+    if(!token) return res.status(401).json("you are unauthorized")
+        jwt.verify(token, "jwt",(err, userid)=>{})
+
+
+
 }
 export const updatePost=(req,res)=>{}
+
+
+
+
 export const deletePost=(req,res)=>{
-    const query="DELETE FROM posts WHERE id=?"
+    const token=req.cookie.auth__token;
+    if(!token) return res.status(401).json("you are unauthorized")
+        jwt.verify(token, "jwt",(err, userid)=>{
+    if(err) return res.status(403).json('tocken is not valid')
+        const query="DELETE FROM posts WHERE id=? AND userid=?";
+    db.query(query,[req.params.id,userid],(err, data)=>{
+        if(err)return res.status(500).json("Error to delete post")
+            return res.status(200).json("post deleted")
+    })
+
+    })
     db.query(query,[req.params.id],(err,data)=>{
         if(err)return res.status(500).json("Error to delete post")
             return res.status(200).json("Post deleted")
